style(position): name fn in hook, move hooks to their own files
